Add products page and inventory dashboard
