+++ frontend/src/components/Sidebar.jsx
@@ -40,8 +40,8 @@ function Sidebar({ user, onLogout }) {
         </svg>
       )
     },
-    { 
-      to: '/inventario', 
+    {
+      to: '/inventario',
       label: 'Inventario',
       icon: (
         <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
@@ -49,17 +49,27 @@ function Sidebar({ user, onLogout }) {
         </svg>
       ),
       submenus: [
-        { 
-          to: '/inventario', 
-          label: 'Productos',
+        {
+          to: '/inventario',
+          label: 'Dashboard',
           icon: (
             <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
               <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M20 7l-8-4-8 4m16 0l-8 4m8-4v10l-8 4m0-10L4 7m8 4v10" />
             </svg>
           )
         },
-        { 
-          to: '/inventario/categorias', 
+        {
+          to: '/inventario/productos',
+          label: 'Productos',
+          icon: (
+            <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M7 7h.01M7 3h5c.512 0 1.024.195 1.414.586l7 7a2 2 0 010 2.828l-7 7a2 2 0 01-2.828 0l-7-7A1.994 1.994 0 013 12V7a4 4 0 014-4z" />
+            </svg>
+          )
+        }
+        ,
+        {
+          to: '/inventario/categorias',
           label: 'Categorías',
           icon: (
             <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
